pyproject.toml
[tool.poetry]
name = "finazon-grpc-python"
packages = [{include = "finazon_grpc_python"}]
include = ["requirements.txt", "finazon_grpc_python/*.p*"]
version = "1.0.131.1"
description = "Finazon gRPC client library for Python"
authors = ["Finazon <team@finazon.io>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/finazon-io/finazon-grpc-python"
repository = "https://github.com/finazon-io/finazon-grpc-python"

[tool.poetry.dependencies]
python = "^3.8"
grpcio-tools = "1.58.0"
grpcio = "1.58.0"
mypy-protobuf = "^3.5.0"


[tool.poetry.group.dev.dependencies]
jinja2 = "^3.1.3"
inflection = "^0.5.1"
autopep8 = "^2.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
